refractor main.py to split commands
